mpl: move shape abstractions to their own translation unit
