+++ src/mpl/src/shapes.cpp
@@ -1,0 +1,32 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2021-2025, The OpenROAD Authors
+
+#include "shapes.h"
+
+namespace mpl {
+
+/*
+  Tiling Operators
+
+  The original implentation used std::pair to represent tilings.
+  These overloads mimic the behavior w.r.t how tilings would get
+  ordered in a std::set.
+*/
+
+bool Tiling::operator==(const Tiling& tiling) const
+{
+  return width_ == tiling.width() && height_ == tiling.height();
+}
+
+bool Tiling::operator<(const Tiling& tiling) const
+{
+  if (width_ != tiling.width()) {
+    return width_ < tiling.width();
+  }
+
+  return height_ < tiling.height();
+}
+
+}
+
+
